refactor colorpicker
